improvements
created a new plane object because the tinker object was gone and made an tinker2.txt because tinker.txt has a wrong rotation.

added the plane.h class to main and made an default constructor for this.

I also made some improvements to main like improving the coded and wrote an function for calculating the mach number so the plane wont go faster than mach 0.8